Place file-node edge tests in shared modules
Move the new backend-agnostic file-node scenario bodies into the shared test
module and expose the requested SQLite wrappers from the parent DB test
module.

Split the older shared scenarios into a small core helper module and move
SQLite-only wrappers into the SQLite helper module so all touched Rust files
stay within the line budget while preserving Postgres coverage.
